Added polygon files and tests to appropriate folders. Modified Makefile.
Polygon files were added to the polygon subfolder of propcov-cpp. Tests
were added to tests-cpp. Makefile to run the tests has not yet been
updated.

A slight modification was made to the Projector class so that it could
compile. The name of a function of one of its dependencies had been
modified without the corresponding update to the Projector class.